moved empty bucket adding functionality out to a util file and generalized
